server/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
flask-migrate
python-dotenv
pinecone-client
langchain
langchain-google-genai
langchain-pinecone
# google-genai
werkzeug
sentence-transformers
psycopg2-binary
gunicorn
pymongo
pydantic